requirements.txt
aerich==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
aiofiles==22.1.0 ; python_version >= "3.10" and python_version < "4.0"
aiogram @ git+https://github.com/aiogram/aiogram.git@fad45c66aa7833eab8c60d87cb2567c4012de655 ; python_version >= "3.10" and python_version < "4.0"
aiogram3-form==0.3.1 ; python_version >= "3.10" and python_version < "4.0"
aiohttp==3.8.4 ; python_version >= "3.10" and python_version < "4.0"
aioschedule==0.5.2 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
dictdiffer==0.9.0 ; python_version >= "3.10" and python_version < "4.0"
et-xmlfile==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
iso8601==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
magic-filter==1.0.9 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "4.0"
openpyxl==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.4 ; python_version >= "3.10" and python_version < "4.0"
pydantic[dotenv]==1.10.4 ; python_version >= "3.10" and python_version < "4.0"
pypika-tortoise==0.1.6 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.10" and python_version < "4.0"
tomlkit==0.11.7 ; python_version >= "3.10" and python_version < "4.0"
tortoise-orm==0.19.2 ; python_version >= "3.10" and python_version < "4.0"
typer==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.8.2 ; python_version >= "3.10" and python_version < "4.0"
